UI: Make milestonegraph re-usable
Change-type: patch
